requirements.txt
taipy==3.0.0
google-auth==2.27.0
google-api-python-client==2.114.0
duckdb==0.9.2
pandas>=2.0.0
google-cloud-storage==2.14.0